pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = ""
authors = ["Sven Grunewaldt <strayer@olle-orks.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
loguru = "^0.7.0"
hikari = "^2.0.0.dev113"
hikari-lightbulb = "^2.3.0"
rq = "^1.10.1"
openai = "^1.12.0"
sentry-sdk = {extras = ["rq"], version = "^1.40.4"}
docker = "^7.0.0"
backoff = "^2.2.1"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]
black = "^23.1a1"
pylint = "^2.12.2"
bandit = "^1.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
